Move ButtonConfig module to correct location
